Add files via upload
In a course I am taking on full stack development in Django, in the JavaScript section the person teaching the course gave us a problem: A spy agency needed a way to verify that the user was actually apart of the agency. So I had to try and do that. I think it turned out pretty well.